+++ CHANGELOG.md
@@ -9,6 +9,7 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 
 ### Added
 - Contribution guidelines
+- Support for attaching URLs to reports
 
 ## [0.12.300] - 2015-09-24
 ### Changed